Refactor hereditary substitutions completely and cleanup dependencies.
Hereditary substitutions now consist of two main components:

 - an inductive datatype SVSub for single-variable substitutions that
   has constructors mirroring the structure of Fin/Ctx/ℕ (to avoid
   green slime) together with an inductive datatype SVRes which
   represents the term or variable resulting from the application of
   an SVSub to a given variable, and a collection of inductively
   defined predicates (Hit, Miss and Hit?) that relate the results of
   such applications/lookups to terms and variables;

 - a set of recursive functions that take a term, a single-variable
   substitution and a simple kind and perform hereditary substitution.

Note that single-variable substitutions are NOT indexed by a simple
kind (unlike the old HSub datatype).  Instead the kind becomes an
explicit parameter of the hsubst functions.  There are also several
judgments representing (simply) kinded hereditary substitutions and
their results, which ARE indexed by the expected kind of the
substitute to make sure the kind parameter of the hsubst is consistent
with the actual kind of the term to be substituted.

Overall, this refactoring reduces the amount of green slime and
explicit fiddling with contexts and variable names (i.e. Fin
instances) though some remains.